+++ packages/expo-splash-screen/src/index.native.ts
@@ -4,8 +4,17 @@ import { SplashScreenNativeModule, SplashScreenOptions } from './SplashScreen.ty
 
 const SplashModule = requireOptionalNativeModule<SplashScreenNativeModule>('ExpoSplashScreen');
 
+const isExpoGo = typeof expo !== 'undefined' && globalThis.expo?.modules?.ExpoGo;
+
 export function setOptions(options: SplashScreenOptions) {
   if (!SplashModule) {
+    return;
+  }
+
+  if (isExpoGo) {
+    console.warn(
+      "'Splashscreen.setOptions' cannot be used in Expo Go. To customize the splash screen, you can use development builds."
+    );
     return;
   }